Fix most -Wall warnings (#115)
* started on basic type errors

* Done. For now

* remove libgfxd binary

* Most warnings fixed

* Fix conflicts and new warnings

* Makefile

* Sig

* Clang format

* destructor

* Add a bunch of destructors

* clear the scalars vector after freeing the elements

* PR fixes (Anghelo Alf)

* Fix most clang -Wall warnings

+++ ZAPD/ZBackground.cpp
@@ -16,7 +16,7 @@ ZBackground::ZBackground(ZFile* nParent) : ZResource(nParent)
 }
 
 ZBackground::ZBackground(const std::string& prefix, const std::vector<uint8_t>& nRawData,
-                         int nRawDataIndex, ZFile* nParent)
+                         uint32_t nRawDataIndex, ZFile* nParent)
 	: ZResource(nParent)
 {
 	rawData.assign(nRawData.begin(), nRawData.end());
@@ -62,7 +62,7 @@ void ZBackground::ParseBinaryFile(const std::string& inFolder, bool appendOutNam
 }
 
 void ZBackground::ExtractFromXML(tinyxml2::XMLElement* reader, const std::vector<uint8_t>& nRawData,
-                                 int nRawDataIndex, const std::string& nRelPath)
+                                 uint32_t nRawDataIndex, const std::string& nRelPath)
 {
 	ZResource::ExtractFromXML(reader, nRawData, nRawDataIndex, nRelPath);
 	DeclareVar("", "");
@@ -123,12 +123,12 @@ void ZBackground::CheckValidJpeg(const std::string& filepath)
 		        "ZBackground::CheckValidJpeg: Warning.\n"
 		        "\t The image is bigger than the screen buffer. File: '%s'.\n"
 		        "\t Image size: %zu bytes.\n"
-		        "\t Screen buffer size: %i bytes.\n",
+		        "\t Screen buffer size: %zu bytes.\n",
 		        filename.c_str(), data.size(), GetRawDataSize());
 	}
 }
 
-int ZBackground::GetRawDataSize()
+size_t ZBackground::GetRawDataSize()
 {
 	// Jpgs use the whole sceen buffer, which is a u16 matrix.
 	return Globals::Instance->cfg.bgScreenHeight * Globals::Instance->cfg.bgScreenWidth * 2;
